Initial commit
ui-curses.scm mostly translated from cmus's ui_curses.c, so it's not
very scheme-y.  This will probably (hopefully) change.  All that's done
right now is to initialize ncurses and handle a few command line
arguments.

mpd-client.scm is a de-modulized and renamed version of the mpd-client
egg.  This is a small enough library to be compiled statically, and
it'll probably be necessary to make changes to it at some point.
Functions in this file are renamed as they're likely to conflict with
functions defined in scmus-client.scm.