'editable' record and text editing functions
The file 'editable.scm' defines the 'editable' record type, which is a
representation of text + a cursor.  Functions are defined to move the
cursor, delete characters before/under the cursor, add characters, etc.

The command-line functions in command-mode.scm are now thin wrappers for
the editable functions.

The run-command function now does a read-eval-print, using the
curses-print function from ui-curses.scm.  This just prints at (0,0).