Initial format string implementation
The format string language is similar to that of cmus, except that ~ is
used instead of % for format specifiers.  Currently the format
specifiers produce substitutions like "<artist>", because there is no
track-info data in scmus yet.  Right-alignment is handled by returning a
pair from scmus-format.

Only single character specifiers are supported so far, so no width,
color, etc..