Screen update system, more tag support
ui-curses.scm defines the function ui-element-changed! which takes as an
argument a symbol naming the ui element which needs to be redrawn.  This
should prevent the proliferation of *-changed global flags, and keep the
mutation contained to ui-curses.scm.

Also, more tags are supported in the format string language, and the
track-info alists also contain more information.