Option system
Options are stored in an alist in option.scm.  The functions get-option
and set-option! do the obvious thing, except that getters and setters
can be defined on an option-by-option basis, so that
get-option/set-option! may do some extra processing.  So far there are
options 'format-current and 'format-status, which contain the format
strings for the current and status lines, respectively.