Key-bind system
This system is very similar to cmus's, except that sequences of keys may
be bound.  For example, to bind "a s d" to (scmus-next!) in the "common"
context:

    (make-binding! '("a" "s" "d") 'common (lambda () (scmus-next!)))

To then remove this binding:

    (unbind! '("a" "s" "d") 'common)

This will also remove any conflicting bindings (bindings with the prefix
"a s d" or bindings for "a" or "a s"), so the above expressions can be
combined in order to forcibly bind keys.