Window abstraction
A window is a view of a list, with a cursor.  That is, a window
describes a sublist of some list, where one element in the sublist is
special (the element under the cursor).  The cursor may be moved.  If
the cursor moves off the edge of the sublist, the window scrolls.

The queue is now described by a window.  The win-move! function is used
to move the cursor in the current window.

Additionally, ~/.scmusrc is loaded into the user's eval environment, so
it may use only the functions available in that environment.