Library view + initial unicode support
The library view lists all the artists in the database.  Currently there
is no behaviour associated with selecting an artist in the library view.

Unicode characters are displayed correctly.  This involved linking with
-lncursesw (instead of -lncurses) and doing a couple setlocale() calls
(which I don't understand in the slightest, but cmus does it and it
works).  There are almost certainly still unicode issues, e.g. calls to
string-length and other non-unicode-aware string functions.