win-deactivate! function
My intention is to use this to navigate backward in the library view.
So the user would activate an artist, which would take them to a list of
that artist's albums.  Then win-deactivate! (with any selection) would
take them back to the list of artists.  This commit only contains the
infrastructure to make this possible.