Multi-level library view
When an artist is selected in the library window, the window changes its
content to display the albums for that artist.  When an album is
selected, a list of dummy entries is displayed (more needs to be done in
the libmpdclient wrapper to get song metadata).  "Deactivating" the
library window takes you back to the previous list (albums, or artists).