current-track, win-selected, track-* functions
current-track returns a track object which can be interrogated with
various track-* functions (e.g. track-artist).  win-selected returns a
track object as well if the current window is a track window.  In the
library artists/albums windows it returns a string (the selected
artist/album name).

With this it should be possible to write e.g. tag editing scripts which
call external tag editors using shell!.