mpd db search
The search windows is split vertically into two panes: the top for
entering queries, and the bottom for displaying results.  Queries are
strings of the form "[tag]:[value]".  Search fields can be added and
removed with the win-add!/win-remove! functions.  win-activate! performs
the search with the given query.

Search fields are edited by running the win-edit! function with a search
field selected.  By default, win-edit! is bound to the "i" key.