Library view rewrite
Full rewrite of library.scm (now library-view.scm).

The main change is that window objects are created as needed in the new
code, as opposed to maintaining a small collection of windows and
reusing them.

Playlists are listed in the toplevel library window, though it is not
possible to view their contents (yet).

When a track is activated (by win-activate!), the track's metadata is
shown.