Expose mpd hostname and port as options
The mpd-address and mpd-port options hold the (default) mpd server
details.  The init-client function is replaced by scmus-connect!, which
uses the default hostname/port if they are not provided as arguments.
scmus-connect! is available in eval-mode as connect!.

This commit also makes it possible to run scmus while not connected to
an mpd server.