Fix connection management bug
This bug occurred when a command was sent, but the reply could not be
read due to an i/o exception (usually a read timeout).  Subsequent
commands would then read from the same connection, getting the response
for an old command.

This is fixed by reconnecting whenever an exception occurs during an mpd
command.

Also, a command line option is added to send an arbitrary command to
mpd.  The response (as a scheme object) is then printed to stdout, and
scmus exits.  This could be useful for piping to other scheme scripts,
or just as an mpc replacement.