Refactor cmdline options code
Code for processing command line options is moved to getopt.scm.  A
record type is used for specifying options rather than lists.
process-opts now returns an alist of options and their values, so that
some processing can be delayed, as needed.

Also, an option is added to specify the MPD password on the command
line.