Incorporate ncurses egg into project
The ncurses egg is compiled statically into scmus.  This way everything
can be linked against libncursesw, which may or may not be significant.

Also, this should make it possible to link statically against ncurses.
(Apparently some systems have only libncurses.a, in which case the
ncurses egg fails to compile.)