Arbitrary metadata + options in format strings
The ~<> specifier may be used to include options (e.g. mpd-address) in
format strings.  If the string inside ~{} is not known to scmus, it is
interpreted as the literal name of a metadata field, which is looked up.

e.g. to get the value of the musicbrainz_trackid metadata:

    ~{musicbrainz_trackid}

to print hostname:port of MPD:

    ~<mpd-address>:~<mpd-port>