Color specifiers in format strings
The syntax is ~<color> where "color" is an integer between -1 and 255
(inclusive) or a color name.  Currently this only sets the foreground
color, but the syntax could be extended to support background colors and
attributes as well (e.g. ~<fg:bg:attr>).

This works by replacing color specifiers with characters in the Unicode
BMP private area.  The function format-addstr does a pass over its
input to interpret these codes.

Color pairs are created dynamically as format-addstr encounters color
codes.  If an appropriate pair already exists, it is used.  Otherwise a
new pair is initialized and used.