Store key bindings as expressions
Rather than storing a thunk, store an expression.  This is a
backwards-incompatible change.

Using an expression has some advantages.  The primary advantage is that
expressions can be passed to user-eval (the sandbox egg doesn't like
thunks for some reason--maybe a bug).  Another advantage is that
expressions can be easily converted to strings for display/editing (in
the options menu, say).

Disadvantages are: some small overhead to evaluate the expressions, and
the bind! interface is a bit different.