Make type information globally available
If the "typed" make variable is set to "y" on the command line, then a
file containing all exported type information will be built and
consulted during compilation.  (The reason this is not enabled by
default is that it precludes incremental compilation.)

This generates some spurious warnings about "conflicting types" for
types involving unions.  I suspect this to be a bug on CHICKEN's end.