General-purpose doubly-linked list implementation
Replace the hacky doubly-linked list implementation in command-line.scm
with a proper implementation.

This implementation uses a sentinel head element and is circular.

This is definitely overkill at this point since it is only used for the
command line history.  The ultimate goal is to use this data structure
to replace the use of regular lists for window data.  This will be a
major upgrade since it means we don't have to scan the whole list to get
the selected row (we can just store it as a pointer into the list).

Also in this commit is the beginnings of a test suite.  Currently the
only tests are for the list implementation.