Allow multiple handlers for events
In preparation the implementation of user hooks.

Also, allow events to trigger other events.  In case this leads to an
infinite recursion in handle-events!, a maximum recursion depth is
specified, after which pending events are discarded.

define-event is renamed define-event-handler, since it may be used
multiple times for the same event.