Implement timers in event system
The function register-timer! is used to delay the execution of a thunk
by a number of seconds.  Timers may be either one-shot (the default) or
recurring.

For convenience, there is a register-timer-event! function which takes
the name of an event to trigger rather than a thunk to execute.

The idea here is to implement scmus-update-client using a generic timer
mechanism rather than putting timer code into scmus-client.scm.  Another
planned use is to have a timer which periodically tries to reconnect if
the client is disconnected.  Eventually there should also be a "safe"
timer API for extensions.