Allow users to hook into event system
Add register-event-handler! to the user API.  The user version of the
function only allows registering handlers for a limited subset of
events.

If a user tries to register a handler for an event which they aren't
allowed to handle, register-event-handler! does nothing.  This is by
design: from the user's perspective, it is as if the handler was
registered but the event never occurs.  In this way, we avoid exposing
internal-only events to the user.

Currently, the only event users can handle is the track-changed event.