Use COOPS
Use COOPS instead of define-record/initform in window.scm.

There is one major drawback of COOPS: defmethod doesn't work across
compilation units.  But it should be fine as long as the whole widget
hierarchy is kept in a single unit (possibly split over multiple files
with include).  Another option is to compile the widget code as a
module, but this complicates the build process/installation.