view: replace event-based update mechanism
Remove the window-changed/view-changed screen update mechanism.

Instead, whenever a window is changed, its parent view is marked as
being damaged.  On every iteration of the main loop, the current view is
checked and if it is damaged, then it is redrawn.

This has the benefit that the current view is drawn at most once per
iteration of the main loop.  With the old system, a view-changed event
could potentially be added multiple times per iteration, causing the
view to be drawn more than once.