Always store window data in window object
We need to update the data-len slot whenever the data changes, so if
we're tracking an external list then we might as well keep and update a
reference to it in *window-data, rather than using data-thunk to return
something else.

This change applies to the queue, status and error windows.