Modules: many
Converted/added modules:
    * command-line
    * editable
    * event
    * input
    * iter
    * scmus-base (lib)
    * scmus-error (new)

Some refactoring was required to accomplish this:
    * command-line has a function COMMAND-LINE-GET-STRING which is used
      to enter eval-mode/search mode.  This avoids a dependency on
      eval-mode.
    * ERROR-SET! is renamed to SCMUS-ERROR-SET! and moved to its own
      module.  This is to prevent scmus-base from having dependencies on
      command-line and event.

Also included in this commit is a partially-complete refactoring of the
user scripting API into its own file.  This is not yet a module, or even
a unit (it's included by eval-mode), but will be eventually (when
eval-mode is a unit).

Also, the copyright year has been bumped, and the GPL has been removed
from mpd-client.scm for compatibility with the upstream source.