Rename 'ncurses' module to 'drewt.ncurses'
CHICKEN wants to load the ncurses egg if it's installed, rather than
loading our own module (@#$%!) so rename it.

As far as I can tell there is no way to tell CHICKEN to prefer importing
our own modules, which are compiled into the executable.  So I guess
going forward every module has to be prefixed to avoid conflict with
some random extension someone might decide to write with the same name.