tui: partial redraws
Only redraw damaged widgets, rather than the whole screen.  This
*should* be a huge performance win, but it currently is not because
<window>s are monolothic (we can't redraw only the damaged parts of a
window because its items are not widgets).

In order to implement this, we cache widget geometry in the PRINT-WIDGET
method so that we can reuse it for redrawing damaged widgets.

Also of note here is that we now have a single root widget which
contains the current view.