Work around bug in ncurses and/or chicken
Work around some issues in ncurses and/or chicken involving the
pair_content and getbkgd functions.

The first issue was that scmus was segfaulting when I rebuilt the
binary.  The location of the segfault was in the find-pair function,
specifically in the call to pair_content.  So I rewrote find-pair to use
the *colors* vector instead of iterating through every possible color
pair (which is a better solution anyways).

Then scmus would run, but the colors were wrong.  It turns out that
getbkgd was returning the default terminal attributes no matter what the
current attributes happened to be.  Is this a bug in ncurses, or perhaps
in the FFI glue?  I have no idea, but as it stands on my development
system getbkgd is broken.

So I replaced the use of getbkgd with a global variable to store the
current color pair.