Add logging system
Add a simple logging system.  This is used for errors and verbose
messages in place of VERBOSE-PRINTF and CONSOLE-PRINTF, which are now
removed.

Additionally, CURRENT-OUTPUT-PORT and CURRENT-ERROR-PORT are set to
custom port objects which wrap output to stdout/stdin with the
WITHOUT-CURSES macro.  So it is now possible to use DISPLAY, WRITE, etc.
without messing with ncurses (thus the removal of CONSOLE-PRINTF).