Play from beginning when PLAY! is called
Play from the beginning of the current track, instead of simply
unpausing.  This changes the behaviour of the default "x" keybinding.

PLAY! now also takes an argument, which can be either a track object or
the queue position of a track to play.

The "play" command is added as a wrapper for PLAY!.