Integrate mouse events into the keybind system
Add "keys" for most of the ncurses mouse events.  These keys are
injected into the keybind system at the view level; the tui code knows
nothing about them.

There is still hardcoded handling for left clicks in HANDLE-INPUT method
for <window>.  This allows the user to do something to the selected row,
without having to implement a mechanism for passing event data to user
code.