Use trie instead of hash table to store commands
In order to (efficiently) implement tab-completion for commands, we need
a data structure that can give us a list of keys given a prefix.

Another benefit of using a trie is that we can (efficiently) check if a
string is a unique prefix of a command.  E.g., "pa" could be used as a
short form for "pause".

Unfortunately, CHICKEN's trie egg does not allow getting a list of keys
by prefix, so I've had to implement my own trie module.