Represent command-line mode as a record
Rather than passing a symbol to COMMAND-LINE-GET-STRING, pass a record
encapsulating the prefix, callback and history for the mode.

With this, command-line no longer has to care about which modes are
valid, and how to deal with them.  This will allow more involved
behaviours to be defined per-mode, such as tab-completion.