Add system for command argument completion
Commands may be defined with the DEFINE-COMMAND/COMPLETION macro to
specify tab-completion for arguments.  The completion behaviour is
specified per-argument, and can be given either as a completion
generator function, or as a list of strings enumerating the valid
completions.

For now, only the 'set' command has argument completion implemented.