window: avoid reprinting whole window on move
Only reprint the whole window if scrolling is required; otherwise just
update the previous and currently selected rows.

Also, WINDOW-H-BORDER is removed.  The horizontal border is now built
into the format strings.  This is preferable as it allows the border to
be removed or increased as the user desires.  And padding functionailty
should not be built into individual widgets in the first place: it
should be provided by a decorator widget.