Allow listening for changes to particular options
Revamp the option system to use objects inheriting from <event-source>,
so that they can be listened on.

With this, we no longer have to call GET-OPTION every time we print a
line controlled by a format string.  Instead we store a local copy of
the format and simply read it from a variable.