tui: prevent all double-rendering (probably)
Before reprinting a widget, we check that:
  * the damaged bit is set on the widget
  * none of the widget's parents have their damaged bit set

So in the case where a parent is reprinted before a child, the parent's
call to PRINT-WIDGET! on the child clears the damaged bit, preventing
the child from being reprinted afterwards.

In the case where a child is reprinted before a parent, we notice that
one of its predecessors has its damaged bit set, and abort printing the
widget.