Don't clear screen area before printing
Scrub the screen *after* printing, so that we can do it efficiently.
Clearing before printing effectively halves our performance when
scrolling through a list-box, which is very noticable on slow hardware
(the performance is still bad with this change, just not as bad).

This means that each widget is responsible for clearing any unused
screen space allocated to it.  Currently this is only implemented for
<list-box>, since that is good enough for scmus.