Add "command" procedure
Add the procedure COMMAND which runs a command.  In order for the
argument parsing to make sense, an implicit quoting mode is added to the
command reader which behaves identically to the double-quoting mode
except that double quotes are also implicitly escaped.

This allows the user to put whatever garbage they want into an argument
string, and the reader will read it as a single argument.  (NOTE: it is
important that we use the command reader on the arguments before calling
EVAL-COMMAND, because '$' is read-syntax.)