Use AlbumArtist for top level in library view
Also, when retrieving a list of tracks for an album, search by
AlbumArtist+Album in case of album name collisions.  This is less than
ideal in cases where compilation albums are poorly tagged (AlbumArtist
not set, or differs within the same album).

Also added here is the (undocumented) WIN-SELECTED procedure, which
retrieves the data of the focused widget.